% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefEGValues2018}
\alias{RefEGValues2018}
\title{Reference EG Vaules - 2018}
\format{
A data frame with 6280 rows and 9 variables
\describe{
\item{Taxon}{Genus species}
\item{Exclude}{should species be excluded (Yes/No)}
\item{Hybrid}{EG values}
\item{US}{EG values}
\item{Standard}{EG values}
\item{US_East}{EG values}
\item{US_Gulf}{EG values}
\item{US_West}{EG values}
\item{Oligochaeta}{Is oligochaeta (Yes/No)}
}
}
\usage{
RefEGValues2018
}
\description{
A dataset containing EG Values for use in MAMBE calculations
}
\keyword{datasets}
