% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SalineSitesPelletier2018}
\alias{SalineSitesPelletier2018}
\title{Saline Standards}
\format{
A data frame with 6280 rows and 9 variables
\describe{
\item{StationID}{}
\item{Replicate}{}
\item{AMBI_Score}{}
\item{S}{}
\item{H}{}
\item{SalZone}{}
}
}
\source{
Good-Bad Benchmarks following Pelletier et al. 2018
}
\usage{
SalineSitesPelletier2018
}
\description{
A dataset containing EG Values for use in MAMBE calculations
}
\keyword{datasets}
